refactor: move `RefTrackingFormFields` to use `Islet` architecture
As part of moving to the Islands (`Islet`) architecture, we're migrating
all client side code to use this method (see https://github.com/guardian/gateway/pull/1513).

This commit moves the `RefTrackingFormFields` to this architecture,
so that it's only dynamically loaded on forms/pages that require these
fields.
